+++ core/api_container/server/server.go
@@ -23,6 +23,7 @@ import (
 const (
 	// If no suite registers within this time, the API container will exit with an error
 	suiteRegistrationTimeout = 10 * time.Second
+	grpcServerStopGracePeriod = 10 * time.Second
 )
 
 type ApiContainerServer struct {
@@ -72,7 +73,19 @@ func (server ApiContainerServer) Run() int {
 
 	// Use force-stop rather than graceful stop so that a hung RPC call to the server (e.g. "AddService" that's
 	// pulling a super-huge image) doesn't prevent the API container from shutting down
-	grpcServer.Stop()
+	serverStoppedChan := make(chan interface{})
+	go func() {
+		grpcServer.GracefulStop()
+		serverStoppedChan <- nil
+	}()
+	select {
+	case <- serverStoppedChan:
+		logrus.Info("gRPC server has exited gracefully")
+	case <- time.After(grpcServerStopGracePeriod):
+		logrus.Warnf("gRPC server failed to stop gracefully after %v; hard-stopping now...", grpcServerStopGracePeriod)
+		grpcServer.Stop()
+		logrus.Info("gRPC server was forcefully stopped")
+	}
 
 	if err := <- grpcServerResultChan; err != nil {
 		logrus.Errorf("gRPC server returned an error after it was done serving:")